convert training program
